post deadline bugfix
I firstly added the category filter to a FAB and then moved it to the hamburger menu. After the deadline passed and I submitted the Project I noticed a bug was preventing the catgory filter from working properly. This commit fixes that bug by reverting the code back where the category filter is a FAB.
I would appreciate if you could take this commit into consideration for the project evaluation. It was done 15 minutes after the deadline.